+++ utensor_cgen/backend/__init__.py
@@ -1,1 +1,2 @@
 from .api import BackendManager
+from .base import Backend
